requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.3
dataclasses-json==0.5.8
decorator==5.1.1
dnspython==2.3.0
docx2txt==0.8
executing==1.2.0
fastapi==0.96.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gptcache==0.1.32
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.142
langchainplus-sdk==0.0.10
llama-index==0.6.0a3
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mongoengine==0.27.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
yarl==1.9.2
